feat(parser): add functionality to parse as Excel and add a new `Parser` class
Implement the function `parse_text_as_excel` and remove the error raised (`NotImplementedError`). The current implementation uses the `pandas` library underneath.

The resulting excel contains 3 spreadsheets: **Audit Entries**, **Audit Fields** and **User Groups**:
- **Audit Entries** contains the entries metadata, things like the `time`, `schema`, the user who made the change, the `action` taken and the `state`

- **Audit Fields** contains the fields data and the corresponding `entry_id` and `time`

- **User Groups** contains the `user_name`, `user_login` and the `group_name`s of the users found in the file

Added new `Format` and `Parser` classes. The `Format` class is an enum that represents the currently supported formats to parse the Audit file to, while the `Parser` class acts as a unified interface to all the parser functions, taking the text to parse and the format to parse to.

Other changes:

- Remove old files from `.gitignore`

- Remove `test.py` file as it was used just to test some quick hacks