docs: add `README.pypi.md` for uploading to PyPI
This is to correct the issue with images not showing on PyPI with `README.md`, because of the use of relative paths for images.

Relative paths for images are to be left untouched on `README.md` to be uploaded to GitHub, so if someone forks the project, the images on that fork do not reference the images on the original repo.